processor manager pybind
